pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-lut"
version = "0.1.0"
description = "Inverted keyword-to-file index for coding agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.58.0",
    "python-dotenv>=1.0.0",
    "rich>=13.9.0",
    "tiktoken>=0.8.0",
    "pathspec>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[project.scripts]
codebase = "agent_lut.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_lut"]

[tool.hatch.build.targets.sdist]
include = ["src/agent_lut", "pyproject.toml", "README.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
